Add new sinkhorn method "mixhorn".
The idea is to do a single iteration in log space, whereafter the K matrix is projected back to linear space and it converges with regular sinkhorn.
It is the fastest by some margin on the GPU with float32 but it is less precise compared to sinkhorn_log.
One idea could be to switch back to log space if it has trouble converging.